% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anagotus.data-package.R
\docType{package}
\name{anagotus.data-package}
\alias{anagotus.data}
\alias{anagotus.data-package}
\title{anagotus.data: Anagotus Label Data}
\description{
This package provides label data for Anagotus specimens stored in various institutions. This includes temporal and spatial data where possible.
}
\author{
\strong{Maintainer}: Neil Birrell \email{nbir012@aucklanduni.ac.nz} (\href{https://orcid.org/0000-0002-7961-1626}{ORCID})

}
\keyword{internal}
